requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0
numpy==1.26.4
scipy==1.14.1
joblib==1.4.2
faiss-cpu==1.9.0
httpx==0.28.1
python-dotenv==1.0.1
matplotlib==3.9.3
pandas==2.2.3
